requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.*
google-cloud-storage==2.*
google-auth==2.*
google-auth-oauthlib==1.*
google-auth-httplib2==0.*
pandas==2.*
numpy==1.*
requests==2.*
python-dotenv==1.* 